+++ test/helper/testInstanceChainNode.ts
@@ -1,14 +1,54 @@
 import { type Logger } from 'pino'
 import { injectable } from 'tsyringe'
+import { SubmittableResult } from '@polkadot/api'
+import { SubmittableExtrinsic } from '@polkadot/api/types'
 
 import { type Env } from '../../src/env.js'
 import { ProxyRequest } from '../../src/models/proxy.js'
-import ChainNode from '../../src/lib/chainNode.js'
+import ChainNode, { EventData } from '../../src/lib/chainNode.js'
 
 @injectable()
 export default class ExtendedChainNode extends ChainNode {
   constructor(logger: Logger, env: Env) {
     super(logger, env)
+  }
+
+  async submitRunProcessForProxy(extrinsic: SubmittableExtrinsic<'promise', SubmittableResult>): Promise<void> {
+    try {
+      this.logger.debug('Submitting Transaction %j', extrinsic.hash.toHex())
+      const unsub: () => void = await extrinsic.send((result: SubmittableResult): void => {
+        this.logger.debug('result.status %s', JSON.stringify(result.status))
+
+        const { dispatchError, status } = result
+
+        if (dispatchError) {
+          this.logger.warn('dispatch error %s', dispatchError)
+
+          unsub()
+          if (dispatchError.isModule) {
+            const decoded = this.api.registry.findMetaError(dispatchError.asModule)
+            throw new Error(`Node dispatch error: ${decoded.name}`)
+          }
+
+          throw new Error(`Unknown node dispatch error: ${dispatchError}`)
+        }
+
+        if (status.isFinalized) {
+          const processRanEvent = result.events.find(
+            ({ event: { method } }) => method === 'ProxyAdded' || 'ProxyRemoved'
+          )
+          const data = processRanEvent?.event?.data as EventData
+          // is there anything sensible I can check here?
+          if (!data) {
+            throw new Error('No data returned')
+          }
+
+          unsub()
+        }
+      })
+    } catch (err) {
+      this.logger.warn(`Error in run process transaction: ${err}`)
+    }
   }
 
   async addProxy({ delegatingAlias, proxyAddress, proxyType, delay = 0 }: ProxyRequest) {
